CONFIG: Change line number field in ConfigBaseNode to 32 bit
Very large config files can theoretically be larger than 65535 lines long.
This would cause the line number to wrap to low values. There is no good
reason to use a short int for the line number.

Also update several printf format string to match the new type.

+++ usr/lib/config/configuration.h
@@ -96,7 +96,7 @@ struct ConfigBaseNode {
     struct ConfigBaseNode *next, *prev;
     unsigned int type;
     char *key;
-    uint16_t line;
+    uint32_t line;
     uint16_t flags;
 };
